Remove slotmap for completion handlers and put pointer in user_data
This is mostly some preparation for the future additioin of "first
class" coroutine support in IoQueue and the introduction of an
OperationHandle. It also just makes the code a bit simpler, though it
might be a bit more dangerous as well, because I allocate manually and
pass that pointer into an io_uring hoping I get it back at some point.
The chances are very high I do, but it's still a possible way to leak
memory.