Remove assert from IoResult::error
It turns out it's useful to be able to do that. I want to make sure the
results are handled carefully, but getting an .error() when there is
none in it, is probably not that bad. Definitely not as bad as getting a
result() if there is an error in it.