Introduce CompleterMap
There was something like this in the past (a slot map) that kept
ownership of the completers and user_data would just contain a key
instead of a pointer.
I realized that pointers for completers might get reused and calling
cancel on a very old OperationHandle might actually cancel an operation
that was either already freed (really bad) or might cancel a new
operation with the same completer pointer (pretty bad).
So I had to go back to having mapping between user_data ids and
completers.