Redesign IoQueue (make IO operations eager)
From the beginning I was unsure whether I should initiate the IO
operation when I construct the awaitable (e.g. call io.recv()) or when I
co_await it. I thought that doing it lazily is "cleaner" and considering
that cppcoro does it this way, it seemed like the right thing. But I had
some features on my list that I had trouble getting in because I
couldn't figure out how to make it work nicely.

The things that bothered me were:
- Cancelation: I want to be able to somehow cancel an IO operation
  (without a cancelation token I pass in), but if it hasn't started
  before I co_await it, I can't get a handle on it to cancel through.
  There is no time between co_await-ing and completing the operation!
- Timeouts: For the callback version I could simply add a timeout
  through an OperationHandle - ezpz, makes sense. But with the coroutine
  versions I would have to somehow "buffer" the timeout in the Awaitable
  and add it after I start the operation. I thought that was silly and I
  didn't want to do it, because I didn't want to do it in such a silly
  way.
- Catching a full submission queue. This is pretty much the same problem
  as cancelation. There is no way for me to check whether the operation
  could be submitted, because it hasn't been submitted before I suspend
  the calling coroutine. I wanted to introduce an error and return it as
  a custom errno value, but I think it would be bad design to do it this
  way for couroutine and a different (prettiert) way for callbacks.
- Multishot operations. I think these can make a huge difference but it
  would require me to issue an operation ONCE and then co_await it
  MULTIPLE TIMES. I would have to create an Awaitable and it would have
  to issue on being co_awaited the first time and then not do anything
  on later times. That seemed wrong.

I realized that I could just issue the operation during/before the
construction of the Awaitable, which would solve all these problems. I
could return an OperationHandle, through which I could cancel, which I
could await multiple times, which I could add a timeout too and through
which I could check if the submission was successful. Additionally I
would get rid of a **ton** of code because I don't need three variants
of each syscall, but just one - one that returns an OperationHandle.
This is so much nicer and simpler and more powerful that I don't care if
it's actually the dirty solution.

Also the whole thing is just simpler. The silly template Awaitable thing
doesn't have to store a function and a bunch of parameters in a
std::tuple. I would simply make coroutines the only first-class citizen
and demote callbacks to a second-class citizen. Then I could get rid of
pointer tagging as well.

Despite it being a bit "dirtier" it's also a bit cleaner, because it
doesn't smell like double initialization as much.

The only risk I see in eager operations is that you might create one and
not co_await it. That would be result in a performance impact that is
likely very hard to find. But I also think it's very likely a rare
mistake to make as it looks very wrong or you are doing something
(unnecessarily?) tricky.

After the changes the code is half the size and does more stuff.
A dream!